Done update reset page when search clans, agent store
