+++ src/pages/Home/NewAgent/ClanItem.tsx
@@ -3,9 +3,9 @@ import { AvatarClanByList } from "@components/AvatarContainer"
 import { PlusIcon } from "@components/Icons/Plus"
 import { VideoThumbnailWrapper } from "@components/VideoThumbnailWrapper"
 import { PATH_NAMES } from "@constants/index"
-import { getConfigClanValue } from "@pages/AgentStore/AgentClansStore"
 import { IGroup } from "@pages/ChatPageOld/ChatContainer/LeftBar/useFetchGroups"
 import { useQueryClient } from "@tanstack/react-query"
+import { getConfigClanValue } from "@utils/clanConfig"
 import React from "react"
 import { useLocation, useNavigate, useParams } from "react-router-dom"
 import { twMerge } from "tailwind-merge"
